prov/efa: Move RDM and DGRAM files into separate folders
