+++ prov/efa/src/rdm/rxr_atomic.c
