+++ prov/efa/src/rdm/rxr_env.c
