+++ prov/efa/test/efa_unit_test_common.c
@@ -1,5 +1,5 @@
 #include "efa_unit_tests.h"
-#include "rxr_pkt_type_base.h"
+#include "rdm/rxr_pkt_type_base.h"
 
 struct fi_info *efa_unit_test_alloc_hints(enum fi_ep_type ep_type)
 {
